avent change 3 category to 2
