+++ assets/js/app.js
@@ -5,12 +5,13 @@ window.jQuery = $;
 
 // this "modifies" the jquery module: adding behavior to it
 // the bootstrap module doesn't export/return anything
-
+require('./router');
 require('./jquery-ui.min');
 require('../../public/bundles/pugxautocompleter/js/autocompleter-jqueryui')
 require('./autocompleter');
-require('./category');
+
 require('./rdcity');
+require('./specification');
 require('bootstrap-fileinput');
 require('../../node_modules/bootstrap-fileinput/js/locales/fr');
 require('../../node_modules/bootstrap-fileinput/js/locales/ar');
@@ -18,6 +19,7 @@ require('../../node_modules/bootstrap-fileinput/js/locales/ar');
 require('./ad-images');
 require('bootstrap');
 require('./sidebar');
+require('./category');
 // or you can include specific pieces
 // require('bootstrap/js/dist/tooltip');
 // require('bootstrap/js/dist/popover');
@@ -25,6 +27,9 @@ require('./sidebar');
 $(document).ready(function() {
     $('[data-toggle="popover"]').popover();
 });
+
+
+
