+++ composer.json
@@ -5,6 +5,7 @@
         "php": "^7.1.3",
         "ext-ctype": "*",
         "ext-iconv": "*",
+        "friendsofsymfony/jsrouting-bundle": "^2.3",
         "friendsofsymfony/user-bundle": "~2.0",
         "kartik-v/bootstrap-fileinput": "@dev",
         "pugx/autocompleter-bundle": "^1.4",
@@ -31,11 +32,12 @@
         "symfony/yaml": "4.2.*"
     },
     "require-dev": {
+        "doctrine/doctrine-fixtures-bundle": "^3.1",
+        "swiftmailer/swiftmailer": "^4.3 || ^5.0 || ^6.0",
         "symfony/debug-pack": "*",
         "symfony/maker-bundle": "^1.11",
         "symfony/profiler-pack": "*",
         "symfony/test-pack": "*",
-        "swiftmailer/swiftmailer": "^4.3 || ^5.0 || ^6.0",
         "symfony/web-server-bundle": "4.2.*"
     },
     "config": {
