+++ config/packages/dev/web_profiler.yaml
@@ -1,6 +1,6 @@
 web_profiler:
     toolbar: true
-    intercept_redirects: false
+    intercept_redirects: true
 
 framework:
     profiler: { only_exceptions: false }
