+++ package.json
@@ -19,6 +19,7 @@
         "jquery": "^3.3.1",
         "jquery-autocomplete": "^1.2.8",
         "jquery-ui": "^1.12.1",
-        "popper.js": "^1.14.7"
+        "popper.js": "^1.14.7",
+        "router": "^1.3.3"
     }
 }
