+++ public/js/fos_js_routes.json
@@ -1,0 +1,1 @@
+{"base_url":"","routes":{"select_subcategory":{"tokens":[["text","\/category\/subcategorys"]],"defaults":[],"requirements":[],"hosttokens":[],"methods":["GET","POST"],"schemes":[]},"select_category":{"tokens":[["text","\/category\/categorys"]],"defaults":[],"requirements":[],"hosttokens":[],"methods":["GET","POST"],"schemes":[]}},"prefix":"","host":"localhost","port":"","scheme":"http"}
