+++ src/Controller/CategoryController.php
@@ -9,6 +9,7 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\HttpFoundation\JsonResponse;
 
 /**
  * @Route("/category")
@@ -86,5 +87,34 @@ class CategoryController extends AbstractController
         }
 
         return $this->redirectToRoute('category_index');
+    }
+
+
+    /**
+     * @Route("/subcategorys", name="select_subcategory", methods="GET|POST", options={"expose"=true})
+     */
+    public function subcategoryAction(Request $request)
+    {
+        $generalcategory_id = $request->request->get('generalcategory_id');
+
+        $em = $this->getDoctrine()->getManager();
+
+        $subcategory = $em->getRepository(Category::class)->findCategoryByParentId($generalcategory_id);
+
+        return new JsonResponse($subcategory);
+    }
+
+    /**
+     * @Route("/categorys", name="select_category", methods="GET|POST", options={"expose"=true})
+     */
+    public function categoryAction(Request $request)
+    {
+        $subcategory_id = $request->request->get('subcategory_id');
+
+        $em = $this->getDoctrine()->getManager();
+
+        $category = $em->getRepository(Category::class)->findCategoryByParentId($subcategory_id);
+
+        return new JsonResponse($category);
     }
 }
