+++ src/Form/CategoryType.php
@@ -12,15 +12,26 @@ use Symfony\Component\Form\FormEvents;
 use Symfony\Component\Form\FormInterface;
 use Symfony\Component\Form\Form;
 use Doctrine\ORM\EntityRepository;
+use App\Form\EventListener\AddCategoryFieldSubscriber;
+use App\Form\EventListener\AddGeneralcategoryFieldSubscriber;
+use App\Form\EventListener\AddSubcategoryFieldSubscriber;
+use App\Repository\CategoryRepository;
 
 class CategoryType extends AbstractType
 {
     public function buildForm(FormBuilderInterface $builder, array $options)
     {
-         $builder
+        /* $builder
             ->add('name')
             ->add('parent')
-        ;
+        ;*/
+      /*  $factort = $builder->getFormFactory();
+        $categorySubscriber = new AddCategoryFieldSubscriber($factort);
+        $builder->addEventSubscriber($categorySubscriber);
+        $subcategorySubscriber = new AddSubcategoryFieldSubscriber($factort);
+        $builder->addEventSubscriber($subcategorySubscriber);
+        $generalcategorySubscriber = new AddGeneralcategoryFieldSubscriber($factort);
+        $builder->addEventSubscriber($generalcategorySubscriber);*/
     }
