--- src/Form/SpecificationType.php
@@ -1,131 +1,0 @@
-<?php
-
-namespace App\Form;
-
-use App\Entity\Specification;
-use Symfony\Component\Form\AbstractType;
-use Symfony\Component\Form\FormBuilderInterface;
-use Symfony\Component\OptionsResolver\OptionsResolver;
-
-class SpecificationType extends AbstractType
-{
-    public function buildForm(FormBuilderInterface $builder, array $options)
-    {
-        $builder
-            ->add('mission')
-            ->add('acitvityArea')
-            ->add('fullPartial')
-            ->add('typeOfContract')
-            ->add('experience')
-            ->add('levelOfStudy')
-            ->add('language')
-            ->add('typeOfTranslation')
-            ->add('material')
-            ->add('placeOfLesson')
-            ->add('levelOfStudent')
-            ->add('brand')
-            ->add('color')
-            ->add('fuelType')
-            ->add('model')
-            ->add('changeGear')
-            ->add('manufactureCompany')
-            ->add('generalSituation')
-            ->add('paperSize')
-            ->add('printingType')
-            ->add('printingColor')
-            ->add('keyboardLanguage')
-            ->add('analogDigital')
-            ->add('machinName')
-            ->add('animalColor')
-            ->add('animalSpecies')
-            ->add('filmName')
-            ->add('bookName')
-            ->add('dvdCd')
-            ->add('languageBookFilm')
-            ->add('originCountry')
-            ->add('clothesMaterial')
-            ->add('clothesSize')
-            ->add('coverMaterial')
-            ->add('shape')
-            ->add('gazType')
-            ->add('heating')
-            ->add('heatingType')
-            ->add('classEnergie')
-            ->add('ges')
-            ->add('eventType')
-            ->add('subjectName')
-            ->add('salary')
-            ->add('durationOfLesson')
-            ->add('maxDistance')
-            ->add('manufacturingYear')
-            ->add('maxManufacturingYear')
-            ->add('minManufacturingYear')
-            ->add('numberOfPassengers')
-            ->add('numberOfDoors')
-            ->add('kilometer')
-            ->add('maxKilometer')
-            ->add('minKilometer')
-            ->add('processor')
-            ->add('hddCapacity')
-            ->add('ram')
-            ->add('screenSizeCm')
-            ->add('screenSizeInch')
-            ->add('capacity')
-            ->add('accuracy')
-            ->add('weight')
-            ->add('caliber')
-            ->add('maxCaliber')
-            ->add('minCaliber')
-            ->add('gauge')
-            ->add('sizePerfume')
-            ->add('shoeSize')
-            ->add('parachuteSize')
-            ->add('number')
-            ->add('animalAge')
-            ->add('width')
-            ->add('height')
-            ->add('kidsClothesSize')
-            ->add('numberOfPersson')
-            ->add('length')
-            ->add('diapersSize')
-            ->add('capacitySize')
-            ->add('numberOfDrawer')
-            ->add('numberOfStaging')
-            ->add('numberOfHead')
-            ->add('washingCapacity')
-            ->add('ability')
-            ->add('tiresSize')
-            ->add('floor')
-            ->add('area')
-            ->add('minArea')
-            ->add('maxArea')
-            ->add('roomNumber')
-            ->add('minRoomNumber')
-            ->add('maxRoomNumber')
-            ->add('numberOfFloors')
-            ->add('hdmi')
-            ->add('cdRoom')
-            ->add('wifi')
-            ->add('usb')
-            ->add('treeInOne')
-            ->add('accessories')
-            ->add('withFreezer')
-            ->add('electricHead')
-            ->add('withOven')
-            ->add('covered')
-            ->add('withFurniture')
-            ->add('withGarden')
-            ->add('withVerandah')
-            ->add('withElevator')
-            ->add('date')
-            ->add('city')
-        ;
-    }
-
-    public function configureOptions(OptionsResolver $resolver)
-    {
-        $resolver->setDefaults([
-            'data_class' => Specification::class,
-        ]);
-    }
-}
