+++ templates/ad/index.html.twig
@@ -19,6 +19,7 @@
                 <th>Donate</th>
                 <th>Date</th>
                 <th>WithDriver</th>
+                <th>Mision</th>
                 <th>TypeOfAd</th>
                 <th>actions</th>
             </tr>
@@ -48,6 +49,7 @@
                 <td>{{ ad.donate ? 'Yes' : 'No' }}</td>
                 <td>{{ ad.date ? ad.date|date('Y-m-d H:i:s') : '' }}</td>
                 <td>{{ ad.withDriver ? 'Yes' : 'No' }}</td>
+                <td>{{ ad.mission }}</td>
                 <td>{{ ad.typeOfAd }}</td>
                 <td>
                     <a href="{{ path('ad_show', {'id': ad.id}) }}">show</a>
