+++ templates/ad/new.html.twig
@@ -1,5 +1,8 @@
 {% extends "layout_user.html.twig" %}
 {% block title %}New Ad{% endblock %}
 {% block content %}
-    {{ include('ad/_form.html.twig') }}
+    <div id="soso">
+        {{ include('ad/_form.html.twig') }}
+    </div>
+
 {% endblock %}
