--- templates/specification/_delete_form.html.twig
@@ -1,5 +1,0 @@
-<form method="post" action="{{ path('specification_delete', {'id': specification.id}) }}" onsubmit="return confirm('Are you sure you want to delete this item?');">
-    <input type="hidden" name="_method" value="DELETE">
-    <input type="hidden" name="_token" value="{{ csrf_token('delete' ~ specification.id) }}">
-    <button class="btn">Delete</button>
-</form>
