--- templates/specification/edit.html.twig
@@ -1,13 +1,0 @@
-{% extends 'base.html.twig' %}
-
-{% block title %}Edit Specification{% endblock %}
-
-{% block body %}
-    <h1>Edit Specification</h1>
-
-    {{ include('specification/_form.html.twig', {'button_label': 'Update'}) }}
-
-    <a href="{{ path('specification_index') }}">back to list</a>
-
-    {{ include('specification/_delete_form.html.twig') }}
-{% endblock %}
